Side Nav final , animated articles final
Figured out the blinking glitch on the side nav panel, figured out the correct way to apply animations on the articles 1 by 1 while scrolling, without using display: 'none". So they are taking up the appropriate space on the page even when they are not there yet. Making it possible not to break our scrolling mechanics and a side nav